gen_ai/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "FastMCP server with uv and .env configuration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.0.0rc1,<4",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ddgs>=9.0",
    "fast-flights>=2.2",
    "loguru>=0.7.0",
    "fake_useragent>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["server.py", "llm.py", "llm_client.py", "tools"]

